refactor: exception 추상화 단계 추가 및 생성자 메세지 인자 제거
