--- src/main/java/wooteco/subway/exception/DuplicateLineException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class DuplicateLineException extends RuntimeException {
-
-    public DuplicateLineException(final String errorMessage) {
-        super(errorMessage);
-    }
-}
