--- src/main/java/wooteco/subway/exception/DuplicateStationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class DuplicateStationException extends RuntimeException {
-
-    public DuplicateStationException(final String exceptionMessage) {
-        super(exceptionMessage);
-    }
-}
