--- src/main/java/wooteco/subway/exception/NotFoundLineException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class NotFoundLineException extends RuntimeException {
-
-    public NotFoundLineException(final String errorMessage) {
-        super(errorMessage);
-    }
-}
