--- src/main/java/wooteco/subway/exception/NotFoundStationException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class NotFoundStationException extends RuntimeException {
-
-    public NotFoundStationException(final String errorMessage) {
-        super(errorMessage);
-    }
-}
