+++ src/main/java/wooteco/subway/exception/SubwayException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception;
+
+public class SubwayException extends RuntimeException {
+
+    public SubwayException(String message) {
+        super(message);
+    }
+}
