+++ src/main/java/wooteco/subway/exception/line/NotFoundLineException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.line;
+
+import wooteco.subway.exception.SubwayException;
+
+public class NotFoundLineException extends SubwayException {
+
+    public NotFoundLineException() {
+        super("해당 노선이 존재하지 않습니다.");
+    }
+}
