+++ src/main/java/wooteco/subway/exception/station/DuplicateStationException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.station;
+
+import wooteco.subway.exception.SubwayException;
+
+public class DuplicateStationException extends SubwayException {
+
+    public DuplicateStationException() {
+        super("중복된 역이 존재합니다.");
+    }
+}
